requirements.txt
httpx==0.28.1
python-dotenv==1.2.1
